Add files via upload
Encapsulation is the first of the 3 principles of object-oriented programming. Thanks to encapsulation, the client is prevented from inputting illogical values, thus keeping the output values meaningful and consistent. However, in order to realize the encapsulation principle, the client's dependence on the background code part is minimized as much as possible and direct intervention in object variables is prevented. The client communicates with the object through constructors and methods.